Don't require the Accept-Ranges header
According to the below, if partial content is supported on a server,
this header is not required.

http://tools.ietf.org/html/rfc2616#section-14.5
https://tools.ietf.org/html/rfc7233#section-2.3

So in my interpretation, and according to the robustness principle, a
client should not enforce the presence of the header.

We have encountered problems because of this on Microsoft Azure and
Cachefly, which both seem to support range requests, but not always set
the Accept-Ranges header.